+++ src/trama/__init__.py
@@ -1,4 +1,4 @@
 """Pacote principal da linguagem trama."""
 
 __all__ = ["__version__"]
-__version__ = "1.3.0"
+__version__ = "2.1.2"